core(feat): Cross project general traits (#16)
## Description

This PR introduces cross-project general traits for the APXM system,
defining common contracts that will be implemented across different
components. The changes include:

- Capability Trait: Defines the contract for tools/functions that can be
invoked by the agent, including parameter validation and invocation
- Operation Executor Trait: Establishes the interface for executing AIS
operations with input/output handling
- LLM Backend Trait: Provides the interface for language model backends
with support for both synchronous and streaming generation
- Enhanced Error Types: Improves error handling with serializable error
types and better source location tracking
- Operation Type Extensions: Adds support for operation attributes and
metadata

## Related Issues

<!-- List the issues this PR closes or references. Use keywords like
"Closes", "Fixes", "Refs": -->
- Closes #8 

## Type of Change

<!-- Check the type of change this PR represents: -->
- [x] Feature - New functionality
- [ ] Bugfix - Bug fix
- [x] Refactor - Code refactoring
- [ ] Docs / Tooling - Documentation or tools
- [ ] Performance - Performance improvement
- [ ] Test - Add or modify tests

## Checklist

<!-- Make sure to complete all relevant items before requesting review:
-->
- [x] Tests added/updated
- [x] All tests pass locally
- [x] Code compiles without errors or warnings
- [ ] Documentation updated (if applicable)
- [x] Code follows project conventions
- [x] Commits follow project format
- [x] No commented or debug code without purpose
- [x] Changes reviewed by author

## Affected Areas

<!-- What areas are affected? Check all that apply: -->
- [ ] `area/all`
- [ ] `area/cli`
- [ ] `area/compiler`
- [x] `area/core`
- [ ] `area/runtime`
- [ ] `area/docs`

## Testing

<!-- Describe how this change was tested: -->
- [x] Unit tests
- [ ] Integration tests
- [ ] Manual testing
- [ ] Other: <!-- specify -->

## Breaking Changes

No breaking changes. This is a new feature that adds trait definitions
without changing existing functionality.

## Additional Information

- All traits implement Send + Sync for concurrent access
- Includes comprehensive test suites for all new traits
- Provides builder patterns for request/response types
- Makes ErrorContext serializable for distributed tracing
- Updates SourceLocation fields to use usize instead of u32 for
consistency

+++ crates/apxm-core/src/error/common.rs
@@ -6,6 +6,8 @@ use std::collections::HashMap;
 use std::fmt;
 use std::time::SystemTime;
 
+use serde::{Deserialize, Serialize};
+
 use crate::types::{AISOperationType, Value};
 
 /// Operations correspond to node in the execution DAG, so OpID is the same as NodeId.
@@ -20,14 +22,14 @@ pub struct SourceLocation {
     /// File path or identifier,
     pub file: String,
     /// Line number (1-indexed).
-    pub line: u32,
+    pub line: usize,
     /// Column number (1-indexed).
-    pub column: u32,
+    pub column: usize,
 }
 
 impl SourceLocation {
     /// Creates a new source location.
-    pub fn new(file: String, line: u32, column: u32) -> Self {
+    pub fn new(file: String, line: usize, column: usize) -> Self {
         SourceLocation { file, line, column }
     }
 }
@@ -42,7 +44,7 @@ impl fmt::Display for SourceLocation {
 ///
 /// This struct allow errors to carry contextyal information such as operation IDs,
 /// trace IDs, and user information for better debugging and observability.
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct ErrorContext {
     /// Operation identifier where the error occurred.
     pub operation_id: Option<OpId>,
